@@ -13,6 +13,7 @@ dependencies:
     sdk: flutter
   
   get:
+  flutter_countdown_timer: ^4.1.0
   font_awesome_flutter: ^9.0.0
   bottom_navy_bar:
   flutter_rating_bar: ^4.0.1
@@ -74,8 +75,9 @@ dependencies:
   uuid: ^3.0.6
   app_links: ^3.2.0
   uni_links: ^0.5.1
-  # webview_flutter: ^3.0.4
-
+  qr_flutter: ^4.0.0
+  dotted_line: ^3.1.0
+  status_view: ^0.0.2
 
   
 
