@@ -6,7 +6,7 @@ import '../controllers/login_controller.dart';
 class HomeBinding implements Bindings {
   @override
   void dependencies() {
-    Get.put(() => LoginController(), permanent: true);
-    Get.put(() => HomeBinding());
+    // Get.put(() => LoginController(), permanent: true);
+    Get.put(HomeController());
   }
 }
