@@ -6,7 +6,7 @@ import 'package:travel_hour/controllers/login_controller.dart';
 class LoginBinding implements Bindings {
   @override
   void dependencies() {
-    Get.put(() => LoginController(), permanent: true);
+    // Get.put(() => LoginController(), permanent: true);
     Get.put(() => LoginControllerV2(), permanent: true);
   }
 }
