@@ -35,10 +35,7 @@ class _ExploreState extends State<Explore> with AutomaticKeepAliveClientMixin {
                   Header(),
                   FeaturedQuest(),
                   QuestTypeScroll(),
-                  // RecentPlaces(),
-                  // SpecialStateOne(),
-                  // SpecialStateTwo(),
-                  // RecommendedPlaces()
+               
                 ],
               ),
             ),
@@ -56,7 +53,7 @@ class Header extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     var controller = Get.find<HomeController>();
-    Rx<City>? dropdownValue = Rx<City>(controller.cityList[1]);
+    Rx<City>? dropdownValue = Rx<City>(controller.dropdownValue);
     return Padding(
       padding: const EdgeInsets.only(left: 15, right: 15, top: 30, bottom: 20),
       child: Column(
@@ -88,9 +85,10 @@ class Header extends StatelessWidget {
                           color: Colors.blue[300],
                         ),
                         onChanged: (City? newValue) {
-                          dropdownValue.value = newValue!;
+                          controller.dropdownValue = newValue!;
+                          dropdownValue.value=newValue;
                           //Get Id City for reload List Quest
-                          controller.cityChoice.value = newValue.id;
+                          controller.areaIdChoice.value = newValue.id;
                         },
                         items: controller.cityList
                             .map<DropdownMenuItem<City>>((City value) {
