@@ -6,16 +6,21 @@ import 'package:flutter/services.dart';
 // ignore: depend_on_referenced_packages
 import 'package:flutter_icons/flutter_icons.dart';
 import 'package:get/get.dart';
+import 'package:travel_hour/controllers/history_controller.dart';
 import 'package:travel_hour/pages/guide.dart';
+import 'package:travel_hour/pages/quest_play.dart';
 
 import 'package:travel_hour/pages/profile.dart';
+import 'package:travel_hour/pages/splashV2.dart';
 import '../controllers/home_controller.dart';
 import 'explore.dart';
 import 'history.dart';
+
 class HomePage extends StatelessWidget {
+    HomeController myController =  Get.put(HomeController());
   @override
-  Widget build(BuildContext context) {
-    var myController = Get.find<HomeController>();
+  Widget build(BuildContext context)  {
+
     var _currentIndex = myController.indexHomePage;
     final views = [
       Explore(),
@@ -26,45 +31,52 @@ class HomePage extends StatelessWidget {
     return Scaffold(
         // appBar: AppBar(title: Text("Flutter Demo")),
         body: Obx(() {
-      return views[_currentIndex.value];
-    }), bottomNavigationBar: Obx(
-      () {
-        return BottomNavyBar(
-          selectedIndex: _currentIndex.value,
-          showElevation: true,
-          itemCornerRadius: 24,
-          curve: Curves.easeIn,
-          onItemSelected: _currentIndex,
-          items: <BottomNavyBarItem>[
-            BottomNavyBarItem(
-              icon: Icon(Icons.home),
-              title: Text('Home'),
-              activeColor: Colors.red,
-              textAlign: TextAlign.center,
-            ),
-            BottomNavyBarItem(
-              icon: Icon(Icons.people),
-              title: Text('Users'),
-              activeColor: Colors.purpleAccent,
-              textAlign: TextAlign.center,
-            ),
-            BottomNavyBarItem(
-              icon: Icon(Icons.history_sharp),
-              title: Text(
-                'History ',
-              ),
-              activeColor: Colors.pink,
-              textAlign: TextAlign.center,
-            ),
-            BottomNavyBarItem(
-              icon: Icon(Icons.account_circle),
-              title: Text('Profile'),
-              activeColor: Colors.blue,
-              textAlign: TextAlign.center,
-            ),
-          ],
-        );
-      },
-    ));
+          if(myController.isLoading.value==true){
+            // print("true nef");
+            return SplashStart(content: 'Waiting Loading Data...',);
+          }else
+          return views[_currentIndex.value];
+        }),
+        floatingActionButton: FloatingActionButton(
+          backgroundColor: Colors.redAccent,
+          child: Icon(Icons.play_arrow_sharp),
+          onPressed: () {Get.put(HistoryController());Get.to(QuestsPlayPage());},
+        ),
+        floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
+        bottomNavigationBar: Obx(
+          () {
+            return NavigationBar(
+              backgroundColor: Colors.white,
+              selectedIndex: _currentIndex.value,
+              // showElevation: true,
+              // itemCornerRadius: 24,
+              // curve: Curves.easeIn,
+              onDestinationSelected: _currentIndex,
+              destinations: <NavigationDestination>[
+                NavigationDestination(
+                  icon: Icon(Icons.home),
+                  label: 'Home',
+                ),
+                NavigationDestination(
+                  icon: Icon(Icons.people),
+                  label: 'Users',
+                  // activeColor: Colors.purpleAccent,
+                  // textAlign: TextAlign.center,
+                ),
+                NavigationDestination(
+                  icon: Icon(Icons.history_sharp),
+                  label: 'History ',
+
+                  // activeColor: Colors.pink,
+                  // textAlign: TextAlign.center,
+                ),
+                NavigationDestination(
+                  icon: Icon(Icons.account_circle),
+                  label: 'Profile',
+                ),
+              ],
+            );
+          },
+        ));
   }
 }
