@@ -1,34 +1,34 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/src/foundation/key.dart';
-import 'package:flutter/src/widgets/framework.dart';
-import 'package:get/get.dart';
-import 'package:travel_hour/controllers/play_controller.dart';
+// import 'package:flutter/cupertino.dart';
+// import 'package:flutter/material.dart';
+// import 'package:flutter/src/foundation/key.dart';
+// import 'package:flutter/src/widgets/framework.dart';
+// import 'package:get/get.dart';
+// import 'package:travel_hour/controllers/play_controller.dart';
 
-class IngrogressPage extends GetView<PlayController> {
-  const IngrogressPage({Key? key}) : super(key: key);
+// class IngrogressPage extends GetView<PlayController> {
+//   const IngrogressPage({Key? key}) : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Column(mainAxisAlignment: MainAxisAlignment.center, children: [
-          Text("Right Answer"),
-          ElevatedButton(
-            onPressed: () {
-              Get.close(0);
-            },
-            child: Text('NextQuestItem', style: TextStyle(fontSize: 16)),
-            style: ElevatedButton.styleFrom(
-              padding: const EdgeInsets.only(
-                  left: 40.0, top: 16.0, bottom: 16.0, right: 40.0),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(12), // <-- Radius
-              ),
-            ),
-          ),
-        ]),
-      ),
-    );
-  }
-}
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
+//       body: Center(
+//         child: Column(mainAxisAlignment: MainAxisAlignment.center, children: [
+//           Text("Right Answer"),
+//           ElevatedButton(
+//             onPressed: () {
+//               Get.close(0);
+//             },
+//             child: Text('NextQuestItem', style: TextStyle(fontSize: 16)),
+//             style: ElevatedButton.styleFrom(
+//               padding: const EdgeInsets.only(
+//                   left: 40.0, top: 16.0, bottom: 16.0, right: 40.0),
+//               shape: RoundedRectangleBorder(
+//                 borderRadius: BorderRadius.circular(12), // <-- Radius
+//               ),
+//             ),
+//           ),
+//         ]),
+//       ),
+//     );
+//   }
+// }
