@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:travel_hour/pages/sign_in.dart';
+import 'package:travel_hour/pages/sign_inV2.dart';
 import 'package:travel_hour/utils/next_screen.dart';
 import 'package:easy_localization/easy_localization.dart';
 
@@ -15,7 +16,7 @@ openSignInDialog(context){
             TextButton(
               onPressed: (){
                 Navigator.pop(context);
-                nextScreenPopup(context, SignInPage(tag: 'popup',));
+                nextScreenPopup(context, LoginScreen(tag: 'popup',));
               }, 
               child: Text('sign in').tr()),
 
